Move each state to separate file
